reuse client socket
